feat: implement file likes functionality with cookie storage and update layout for improved responsiveness
